requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=4.0.0
pdfplumber>=0.10.0
pydantic>=2.5.0
chromadb>=0.4.22
numpy>=1.24.0
